LPS-59420 Replace manipulations of tables to use the Services

+++ modules/portal/portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/upgrade/v1_3_0/UpgradeKaleoInstance.java
@@ -14,10 +14,16 @@
 
 package com.liferay.portal.workflow.kaleo.upgrade.v1_3_0;
 
+import com.liferay.dynamic.data.lists.model.DDLRecord;
 import com.liferay.dynamic.data.lists.service.DDLRecordLocalService;
+import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.util.PortalUtil;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 
 /**
  * @author Christopher Kian
@@ -32,17 +38,48 @@ public class UpgradeKaleoInstance extends UpgradeProcess {
 			String tableName, String columnName, Object columnValue)
 		throws Exception {
 
-		StringBundler sb = new StringBundler(7);
+		Connection con = null;
+		PreparedStatement ps = null;
+		ResultSet rs = null;
 
-		sb.append("delete from ");
-		sb.append(tableName);
-		sb.append(" where ");
-		sb.append(columnName);
-		sb.append(" = ");
-		sb.append(columnValue);
-		sb.append(" and classPK not in (select recordId from DDLRecord)");
+		try {
+			con = DataAccess.getUpgradeOptimizedConnection();
 
-		runSQL(sb.toString());
+			StringBundler sb = new StringBundler(6);
+			sb.append("select * from ");
+			sb.append(tableName);
+			sb.append(" where ");
+			sb.append(columnName);
+			sb.append(" = ");
+			sb.append(columnValue);
+
+			ps = con.prepareStatement(sb.toString());
+
+			rs = ps.executeQuery();
+
+			if (rs.next()) {
+				long classPK = rs.getLong("classPK");
+
+					DDLRecord ddlRecord = _ddlRecordLocalService.fetchDDLRecord(
+						classPK);
+
+					if (ddlRecord == null) {
+						sb = new StringBundler(8);
+						sb.append("delete from ");
+						sb.append(tableName);
+						sb.append(" where ");
+						sb.append(columnName);
+						sb.append(" = ");
+						sb.append(columnValue);
+						sb.append(" and classPK = ");
+						sb.append(classPK);
+						runSQL(sb.toString());
+					}
+			}
+		}
+		finally {
+			DataAccess.cleanUp(con, ps, rs);
+		}
 	}
 
 	@Override
